+++ example/lib/main.dart
@@ -1,23 +1,20 @@
 import 'package:code_on_the_rocks/code_on_the_rocks.dart';
 import 'package:flutter/material.dart';
 
-void main() {
-  runApp(const MyApp());
-}
+void main() => runApp(const MyApp());
 
 class MyApp extends StatelessWidget {
   const MyApp({super.key});
 
-  // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'Flutter Demo',
+      title: 'Code on the Rocks',
       theme: ThemeData(
         colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
         useMaterial3: true,
       ),
-      home: HomeView(),
+      home: const HomeView(),
     );
   }
 }
@@ -27,39 +24,95 @@ class HomeView extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return ViewModelBuilder<HomeViewModel>(
-      modelToRegister: HomeViewModel(),
-      builder: (context, model, child) {
-        return  Scaffold(body: Center(
-          child: Column(
-            mainAxisSize: MainAxisSize.min,
+    return HomeViewModelBuilder(
+      builder: (context, model) {
+        return Scaffold(
+          appBar: AppBar(
+            title: Text(model.title),
+          ),
+          body: Stack(
             children: [
-              Text(model.counter.value.toString()),
-              ElevatedButton(onPressed: (){
-                model.setCounter(model.counter.value + 1);
-              }, child: Text('Increment'))
+              Center(
+                child: Column(
+                  mainAxisSize: MainAxisSize.min,
+                  children: [
+                    ValueListenableBuilder(valueListenable: model.counter, builder: (context, value, child) => Text(value.toString())),
+                    const SeparatedCounter(),
+                    ElevatedButton(
+                      onPressed: model.isLoading
+                          ? () {}
+                          : () async {
+                        HomeViewModel().of(context).incrementCounterWithSetState();
+                        model.setLoading(false);
+                      },
+                      child: const Text('Increment using setState'),
+                    ),
+                    ElevatedButton(
+                      onPressed: model.isLoading
+                          ? () {}
+                          : () {
+                        model.incrementCounterWithValueNotifier();
+                      },
+                      child: const Text('Increment using ValueNotifier'),
+                    )
+                  ],
+                ),
+              ),
+              if (model.isLoading) const ColoredBox(color: Colors.black12, child: Center(child: CircularProgressIndicator()))
             ],
           ),
-        ));
+        );
       },
     );
   }
 }
 
-class HomeViewModel extends ChangeNotifier {
-  ValueNotifier<bool> loading = ValueNotifier(false);
+class SeparatedCounter extends StatelessWidget {
+  const SeparatedCounter({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Text(HomeViewModel().of(context).counter.value.toString());
+  }
+}
+
+class HomeViewModelBuilder extends ViewModelBuilder<HomeViewModel> {
+   const HomeViewModelBuilder({
+    super.key,
+    required super.builder,
+  });
+
+  @override
+  State<StatefulWidget> createState() => HomeViewModel();
+}
+
+class HomeViewModel extends ViewModel<HomeViewModel> {
+
+  static HomeViewModel of_(BuildContext context) => (context.dependOnInheritedWidgetOfExactType<ViewModelProvider<ViewModel<HomeViewModel>>>()!.state) as HomeViewModel;
+
+  final String title = 'Home';
 
   ValueNotifier<int> counter = ValueNotifier(0);
 
-  void setCounter(int val){
-    counter.value = val;
-    notifyListeners();
+  Future<void> incrementCounterWithSetState() async {
+    setState(() {
+      counter.value = counter.value + 1;
+    });
   }
 
-  bool get isLoading => loading.value;
+  void incrementCounterWithValueNotifier() {
+    counter.value++;
+  }
 
-  void setLoading(bool val) {
-    loading.value = val;
-    notifyListeners();
+  @override
+  void initState() {
+    debugPrint('Initialize');
+    super.initState();
+  }
+
+  @override
+  void dispose() {
+    debugPrint('Dispose');
+    super.dispose();
   }
 }